+++ NEWS.md
@@ -24,6 +24,9 @@ merging adjacent SBSs to DBS.
 producing memory errors when the input VCF has many variants (e.g. >70000 lines).
 Used `fuzzyjoin::genome_left_join` instead.
 
+* Fixed a bug in internal function `MakeDataFrameFromVCF` when renaming column
+`#CHROM` to `CHROM`.
+
 # 3.0.1
 * Updated two internal functions `GeneratePCAWGAliquotID` and
 `GeneratePCAWGAliquotID2`.